fixed typos
in the figure caption under the first figure, there was text repeated; see if my edit (to remove the duplication) is correct. i'm not sure if the first sentence is meant to be the title of the figure and should maybe be in italics or something.

fixed a few other more minor typos as well.